Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    double n;
-    scanf("%lf",&n);
-    printf("You entered: %lf",n);
+    float n;
+    scanf("%f",&n);
+    printf("You entered: %.5f",n);
     return 0;
 }
